Playing with runtime testing methods
Created a new runtime testing method that uses chrono and timestamps as opposed to clock cycles. This is beneficial for when you are running on a system with variable clock timings. Also made the rdtsc() function not return negative numbers (for now). Need to expand that to all the programs and recompile them.